+++ crates/assay-types/src/worktree.rs
@@ -45,6 +45,7 @@ inventory::submit! {
 ///
 /// Returned by worktree create and list operations.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
+#[serde(deny_unknown_fields)]
 pub struct WorktreeInfo {
     /// The spec slug this worktree is associated with.
     pub spec_slug: String,
@@ -62,6 +63,7 @@ pub struct WorktreeInfo {
 ///
 /// Extends worktree information with dirty state, HEAD commit, and ahead/behind counts.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
+#[serde(deny_unknown_fields)]
 pub struct WorktreeStatus {
     /// The spec slug this worktree is associated with.
     pub spec_slug: String,
@@ -74,9 +76,9 @@ pub struct WorktreeStatus {
     /// Whether the worktree has uncommitted changes.
     pub dirty: bool,
     /// Number of commits ahead of base branch. `None` when the base ref is missing.
-    pub ahead: Option<usize>,
+    pub ahead: Option<u32>,
     /// Number of commits behind base branch. `None` when the base ref is missing.
-    pub behind: Option<usize>,
+    pub behind: Option<u32>,
     /// The base branch this worktree was created from (from metadata).
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub base_branch: Option<String>,
